chore: Move part of error.rs to winit-core
